% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_address.R
\name{read_address}
\alias{read_address}
\title{Adressen einlesen}
\usage{
read_address(path_address)
}
\arguments{
\item{path_address}{Dateipfad für Adressen; Excel-Arbeitsmappe mit xlsm-Erweiterung}

\item{col_types}{Spaltentypen}
}
\description{
read_address() erlaubt das Einlesen der Adressinformationen
}
\examples{
read_address()
}
\keyword{Adressen}
