% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ew.R
\name{read_ew}
\alias{read_ew}
\title{Einwilligungen einlesen}
\usage{
read_ew(path_ew, id_col, datum_col, s_int, m_int, n_survey)
}
\arguments{
\item{path_ew}{Dateipfad für Einwilligungen; tab-delimited csv}

\item{id_col}{Name der ID-Spalte}

\item{datum_col}{Name der Einwilligungsdatumsspalte}

\item{s_int}{Tage zwischen Einwilligung und Einladung}

\item{m_int}{Anzahl der Monate zwischen Einladungen}

\item{n_survey}{Anzahl der Einladungswellen / Surveys}
}
\description{
read_ew() erlaubt das Einlesen von Einwilligungen
}
\examples{
read_ew()
}
\keyword{Einwillungen}
