% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lime_resp.R
\name{read_lime_resp}
\alias{read_lime_resp}
\title{Antwortstatus und -datum von LimeSurvey einlesen}
\usage{
read_lime_resp(user, pw, api, complete, id)
}
\arguments{
\item{user}{LimeSurvey-User}

\item{pw}{LimeSurvey-Passwort}

\item{api}{LimeSurvey-API-Schnittstelle}

\item{complete}{vollständigkeitsstatus der Antworten}

\item{id}{LimeSurvey-ID}
}
\description{
read_lime_resp() erlaubt das Einlesen von Informationen aus LimeSurvey
}
\examples{
read_lime_resp()
}
