% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_invite_3_fun.R
\name{send_invite_3_fun}
\alias{send_invite_3_fun}
\title{Einladungsmails ab 2. Wellen versenden}
\usage{
send_invite_3_fun(dat, senden)
}
\arguments{
\item{dat}{Dataframe mit Adressinformationen der Teilnehmer die angeschrieben werden sollen}

\item{senden}{Sollen mail sofort gesendet werden oder erst in Outlook geöffnet werden (TRUE, FALSE)}
}
\description{
send_invite_2_fun() erlaubt das versenden von Einladungsemails an die Teilnehmer innerhalb ab der 2. Welle
}
\examples{
send_invite_2_fun()
}
