% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_invite_fun.R
\name{send_invite_fun}
\alias{send_invite_fun}
\title{Einladungsmails versenden}
\usage{
send_invite_fun(dat, senden)
}
\arguments{
\item{dat}{Dataframe mit Adressinformationen der Teilnehmer die angeschrieben werden sollen}

\item{senden}{Sollen mail sofort gesendet werden oder erst in Outlook geöffnet werden (TRUE, FALSE)}
}
\description{
send_invite_fun() erlaubt das versenden von ersten Einladungsemails an die Teilnehmer innerhalb der ersten Welle
}
\examples{
send_invite_fun()
}
