% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_remind_2_fun_3.R
\name{send_remind_2_fun_3}
\alias{send_remind_2_fun_3}
\title{Dritten / Letzten Reminder versenden}
\usage{
send_remind_2_fun_3(dat, senden)
}
\arguments{
\item{dat}{Dataframe mit Adressinformationen der Teilnehmer die angeschrieben werden sollen}

\item{senden}{Sollen mail sofort gesendet werden oder erst in Outlook geöffnet werden (TRUE, FALSE)}
}
\description{
send_remind_2_fun_3() erlaubt das versenden der dritten, und letzten, Erinnerungsmail an die Teilnehmer innerhalb der ersten Welle
}
\examples{
send_remind_2_fun_3()
}
