% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_remind_fun_1.R
\name{send_remind_fun_1}
\alias{send_remind_fun_1}
\title{Ersten Reminder versenden}
\usage{
send_remind_fun_1(dat, senden)
}
\arguments{
\item{dat}{Dataframe mit Adressinformationen der Teilnehmer die angeschrieben werden sollen}

\item{senden}{Sollen mails sofort gesendet werden oder erst in Outlook geöffnet werden (TRUE, FALSE)}
}
\description{
send_remind_fun_1() erlaubt das versenden der ersten Erinnerungsmail an die Teilnehmer innerhalb der ersten Welle
}
\examples{
send_remind_fun_1()
}
