% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_remind_fun_2.R
\name{send_remind_fun_2}
\alias{send_remind_fun_2}
\title{Zweiten Reminder versenden}
\usage{
send_remind_fun_2(dat, senden)
}
\arguments{
\item{dat}{Dataframe mit Adressinformationen der Teilnehmer die angeschrieben werden sollen}

\item{senden}{Sollen mail sofort gesendet werden oder erst in Outlook geöffnet werden (TRUE, FALSE)}
}
\description{
send_remind_fun_2() erlaubt das versenden der zweiten Erinnerungsmail an die Teilnehmer innerhalb der ersten Welle
}
\examples{
send_remind_fun_2()
}
