% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_lime_merge.R
\name{sent_lime_merge}
\alias{sent_lime_merge}
\title{Verbinden von LimeSurvey mit bisherigen Teilnehmerantwortdaten}
\usage{
sent_lime_merge(dat_sent, dat_lime, int_len, int_n)
}
\arguments{
\item{dat_sent}{Bisherige eingeladene Teilnehmerdaten, dataframe}

\item{dat_lime}{Limesurvey Antwortdaten, dataframe}

\item{int_len}{Länge des Intervalls, in Tagen, zwischen heutigen und Limesurvey-Antwortdatum}

\item{int_n}{Maximale Anzahl der Reminder}
}
\description{
sent_lime_merge() erlaubt das Verbinden bisherigen eingeladenden Teilnehmerinformationen von Informationen aus LimeSurvey und Aktualisierung des Erinnerungsstatus
}
\examples{
sent_lime_merge()
}
