% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_welle_on_db.R
\name{update_welle_on_db}
\alias{update_welle_on_db}
\title{Update der Befragungsspalte auf der Teilnehner-Datenbank}
\usage{
update_welle_on_db(befragung_code, token)
}
\arguments{
\item{befragung_code}{Befragungswelle}

\item{token}{Token/L3-ID des Befragungswelle die geupdatet werden soll}
}
\description{
update_welle_on_db() erlaubt ein Update der Befragungsspalte auf der Teilnehner-Datenbank
}
\examples{
update_welle_on_db()
}
