% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widerruf_to_db.R
\name{widerruf_to_db}
\alias{widerruf_to_db}
\title{Widerrufe auf die Teilnehmerdatenbank schreiben}
\usage{
widerruf_to_db(wid_code, token)
}
\arguments{
\item{wid_code}{Kodierung von Widerrufen auf den Teilnehmerdatenbank}

\item{token}{Token/L3-ID des Widerufs}
}
\description{
widerruf_to_db() erlaubt das Scheiben von Widerrufen auf die Teilnehmerdatenbank
}
\examples{
widerruf_to_db()
}
