% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tn_to_db.R
\name{write_tn_to_db}
\alias{write_tn_to_db}
\title{Neue Token auf Teilnehmerdatenbank schreiben}
\usage{
write_tn_to_db(user, pw, dbname, host, port, tab_name, tn_to_add)
}
\arguments{
\item{user}{DB-User}

\item{pw}{DB-Passwort}

\item{dbname}{DB-Name}

\item{host}{DB-Host}

\item{port}{DB-Port}

\item{tab_name}{Name der DB-Tabelle}

\item{tn_to_add}{Tabelle mit Teilnehmer ID die an die Datenbank drangehangen werden soll}
}
\description{
write_tn_to_db() erlaubt das Schreiben neuen Token/L3-IDs auf die Teilnehmerdatenbank
}
\examples{
write_tn_to_db()
}
